chore: move security configuration resolution
