+++ src/client/profile-provider.ts
@@ -3,11 +3,6 @@ import {
   assertProfileDocumentNode,
   assertProviderJson,
   FILE_URI_PROTOCOL,
-  HttpScheme,
-  isApiKeySecurityValues,
-  isBasicAuthSecurityValues,
-  isBearerTokenSecurityValues,
-  isDigestSecurityValues,
   isFileURIString,
   isMapFile,
   isProfileFile,
@@ -16,8 +11,6 @@ import {
   prepareProviderParameters,
   ProfileDocumentNode,
   ProviderJson,
-  SecurityScheme,
-  SecurityType,
   SecurityValues,
 } from '@superfaceai/ast';
 import createDebug from 'debug';
@@ -28,11 +21,9 @@ import { Config } from '../config';
 import { UnexpectedError } from '../internal';
 import {
   invalidProfileError,
-  invalidSecurityValuesError,
   localProviderAndRemoteMapError,
   providersDoNotMatchError,
   referencedFileNotFoundError,
-  securityNotFoundError,
 } from '../internal/errors.helpers';
 import {
   MapInterpreter,
@@ -61,6 +52,7 @@ import { MapInterpreterEventAdapter } from './failure/map-interpreter-adapter';
 import { ProfileConfiguration } from './profile';
 import { ProviderConfiguration } from './provider';
 import { fetchBind, fetchMapSource, fetchProviderInfo } from './registry';
+import { resolveSecurityConfiguration } from './security';
 
 function forceCast<T>(_: unknown): asserts _ is T {}
 
@@ -341,7 +333,7 @@ export class ProfileProvider {
       );
     }
 
-    const securityConfiguration = this.resolveSecurityConfiguration(
+    const securityConfiguration = resolveSecurityConfiguration(
       providerInfo.securitySchemes ?? [],
       securityValues,
       providerName
@@ -697,88 +689,5 @@ export class ProfileProvider {
     }
 
     return resolved;
-  }
-
-  private resolveSecurityConfiguration(
-    schemes: SecurityScheme[],
-    values: SecurityValues[],
-    providerName: string
-  ): SecurityConfiguration[] {
-    const result: SecurityConfiguration[] = [];
-
-    for (const vals of values) {
-      const scheme = schemes.find(scheme => scheme.id === vals.id);
-      if (scheme === undefined) {
-        const definedSchemes = schemes.map(s => s.id);
-        throw securityNotFoundError(providerName, definedSchemes, vals);
-      }
-
-      const invalidSchemeValuesErrorBuilder = (
-        scheme: SecurityScheme,
-        values: SecurityValues,
-        requiredKeys: [string, ...string[]]
-      ) => {
-        const valueKeys = Object.keys(values).filter(k => k !== 'id');
-
-        return invalidSecurityValuesError(
-          providerName,
-          scheme.type,
-          scheme.id,
-          valueKeys,
-          requiredKeys
-        );
-      };
-
-      if (scheme.type === SecurityType.APIKEY) {
-        if (!isApiKeySecurityValues(vals)) {
-          throw invalidSchemeValuesErrorBuilder(scheme, vals, ['apikey']);
-        }
-
-        result.push({
-          ...scheme,
-          ...vals,
-        });
-      } else {
-        switch (scheme.scheme) {
-          case HttpScheme.BASIC:
-            if (!isBasicAuthSecurityValues(vals)) {
-              throw invalidSchemeValuesErrorBuilder(scheme, vals, [
-                'username',
-                'password',
-              ]);
-            }
-
-            result.push({
-              ...scheme,
-              ...vals,
-            });
-            break;
-
-          case HttpScheme.BEARER:
-            if (!isBearerTokenSecurityValues(vals)) {
-              throw invalidSchemeValuesErrorBuilder(scheme, vals, ['token']);
-            }
-
-            result.push({
-              ...scheme,
-              ...vals,
-            });
-            break;
-
-          case HttpScheme.DIGEST:
-            if (!isDigestSecurityValues(vals)) {
-              throw invalidSchemeValuesErrorBuilder(scheme, vals, ['digest']);
-            }
-
-            result.push({
-              ...scheme,
-              ...vals,
-            });
-            break;
-        }
-      }
-    }
-
-    return result;
   }
 }
